add bleaching summary as reusuable componnent
