+++ src/App/mermaidData/mermaidDataProptypes.js
@@ -287,6 +287,20 @@ const _reefSlopePropType = PropTypes.shape({
   ),
 })
 
+export const obsColoniesBleachedPropType = PropTypes.arrayOf(
+  PropTypes.shape({
+    count_20: PropTypes.number,
+    count_50: PropTypes.number,
+    count_80: PropTypes.number,
+    attribute: PropTypes.string,
+    count_100: PropTypes.number,
+    count_dead: PropTypes.number,
+    count_pale: PropTypes.number,
+    growth_form: PropTypes.string,
+    count_normal: PropTypes.number,
+  }),
+)
+
 export const choicesPropType = PropTypes.shape({
   fishsizebins: _fishSizeBinPropType,
   belttransectwidths: _beltTransectWidthPropType,
