+++ src/components/pages/submittedRecordPages/SubmittedBleaching/SubmittedBleachingObservationTable.js
@@ -4,17 +4,13 @@ import {
   submittedBleachingPropType,
 } from '../../../../App/mermaidData/mermaidDataProptypes'
 import { inputOptionsPropTypes } from '../../../../library/miscPropTypes'
-import { SubmittedObservationStickyTable, Tr, Td, Th } from '../../../generic/Table/table'
-import {
-  TheadItem,
-  FormSubTitle,
-  ObservationsSummaryStats,
-  UnderTableRow,
-} from '../SubmittedFormPage.styles'
+import { SubmittedObservationStickyTable, Tr, Td } from '../../../generic/Table/table'
+import { TheadItem, FormSubTitle } from '../SubmittedFormPage.styles'
 import { InputWrapper } from '../../../generic/form'
 import { StyledOverflowWrapper } from '../../collectRecordFormPages/CollectingFormPage.Styles'
 import { getObjectById } from '../../../../library/getObjectById'
 import { getOptions } from '../../../../library/getOptions'
+import BleachingObservationSummaryStats from '../../../BleachingObservationSummaryStats/BleachingObservationSummaryStats'
 
 const SubmittedBleachingObservationTable = ({
   benthicAttributeOptions,
@@ -39,45 +35,6 @@ const SubmittedBleachingObservationTable = ({
     </Tr>
   ))
 
-  const getTotalOfColonies = () => {
-    const totals = obs_colonies_bleached.map(
-      (item) =>
-        item.count_20 +
-        item.count_50 +
-        item.count_80 +
-        item.count_100 +
-        item.count_dead +
-        item.count_normal +
-        item.count_pale,
-    )
-
-    return totals.reduce((acc, currentVal) => acc + currentVal, 0)
-  }
-
-  const getTotalOfCoralGenera = () => {
-    const attributeIds = obs_colonies_bleached.map((item) => item.attribute)
-    const uniqueAttributeIds = [...new Set(attributeIds)]
-
-    return uniqueAttributeIds.length
-  }
-
-  const getPercentageOfColonies = (colonyType) => {
-    let totals = 0
-
-    if (colonyType === 'bleached') {
-      totals = obs_colonies_bleached.map(
-        (item) => item.count_20 + item.count_50 + item.count_80 + item.count_100 + item.count_dead,
-      )
-    } else {
-      totals = obs_colonies_bleached.map((item) => item[colonyType])
-    }
-
-    return (
-      (totals.reduce((acc, currentVal) => acc + currentVal, 0) / getTotalOfColonies()) *
-      100
-    ).toFixed(1)
-  }
-
   return (
     <InputWrapper>
       <FormSubTitle id="table-label">Observations</FormSubTitle>
@@ -100,32 +57,7 @@ const SubmittedBleachingObservationTable = ({
           <tbody>{observationsBleaching}</tbody>
         </SubmittedObservationStickyTable>
       </StyledOverflowWrapper>
-      <UnderTableRow>
-        <ObservationsSummaryStats>
-          <tbody>
-            <Tr>
-              <Th>Total number of colonies</Th>
-              <Td>{getTotalOfColonies()}</Td>
-            </Tr>
-            <Tr>
-              <Th>Total number of coral genera</Th>
-              <Td>{getTotalOfCoralGenera()}</Td>
-            </Tr>
-            <Tr>
-              <Th>% Normal colonies</Th>
-              <Td>{getPercentageOfColonies('count_normal')}</Td>
-            </Tr>
-            <Tr>
-              <Th>% Pale colonies</Th>
-              <Td>{getPercentageOfColonies('count_pale')}</Td>
-            </Tr>
-            <Tr>
-              <Th>% Bleached colonies</Th>
-              <Td>{getPercentageOfColonies('bleached')}</Td>
-            </Tr>
-          </tbody>
-        </ObservationsSummaryStats>
-      </UnderTableRow>
+      <BleachingObservationSummaryStats obsColoniesBleached={obs_colonies_bleached} />
     </InputWrapper>
   )
 }
